+++ components/core/ui/searchbar/src/main/java/com/flipperdevices/core/ui/searchbar/ComposableSearchBar.kt
@@ -14,6 +14,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import com.flipperdevices.core.ui.ktx.SetUpStatusBarColor
 import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.res.R as DesignSystem
@@ -43,6 +44,7 @@ private fun ComposableSearchBarInternal(
     onChangeText: (String) -> Unit,
     onBack: () -> Unit
 ) {
+    SetUpStatusBarColor(LocalPallet.current.background)
     Row(
         modifier = Modifier.background(LocalPallet.current.background),
         verticalAlignment = Alignment.CenterVertically