+++ components/deeplink/api/src/main/java/com/flipperdevices/deeplink/model/Deeplink.kt
@@ -1,35 +1,36 @@
 package com.flipperdevices.deeplink.model
 
-import android.content.Intent
 import android.os.Parcelable
 import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
 import kotlinx.parcelize.Parcelize
+import kotlinx.serialization.Serializable
 
-sealed class Deeplink(
-    val isInternal: Boolean = false,
-    open var intent: Intent? = null
-) : Parcelable {
+@Serializable
+sealed class Deeplink : Parcelable {
     @Parcelize
+    @Serializable
     data class FlipperKey(
         val path: FlipperFilePath? = null,
         val content: DeeplinkContent? = null
     ) : Deeplink()
 
     @Parcelize
+    @Serializable
     data class WidgetOptions(
         val appWidgetId: Int
     ) : Deeplink()
 
     @Parcelize
+    @Serializable
     data class OpenKey(
         val keyPath: FlipperKeyPath
     ) : Deeplink()
 
     @Parcelize
+    @Serializable
     data class WebUpdate(
         val url: String,
         val name: String,
-        override var intent: Intent?
-    ) : Deeplink(isInternal = true, intent = intent)
+    ) : Deeplink()
 }