+++ components/deeplink/impl/src/main/java/com/flipperdevices/deeplink/impl/parser/delegates/DeepLinkWebUpdater.kt
@@ -7,11 +7,10 @@ import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.deeplink.api.DeepLinkParserDelegate
 import com.flipperdevices.deeplink.model.DeepLinkParserDelegatePriority
 import com.flipperdevices.deeplink.model.Deeplink
-import com.flipperdevices.deeplink.model.DeeplinkConstants
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
 
-private val HOSTS = listOf(DeeplinkConstants.LAB_FLIPPER_NET, DeeplinkConstants.MY_FLIPPER_DEV)
+private val SUPPORTED_HOSTS = listOf("lab.flipper.net", "my.flipp.dev")
 private const val QUERY_URL = "url"
 private const val QUERY_VERSION = "version"
 private const val QUERY_CHANNEL = "channel"
@@ -27,7 +26,7 @@ class DeepLinkWebUpdater @Inject constructor() : DeepLinkParserDelegate, LogTagP
         if (intent.data == null) {
             return DeepLinkParserDelegatePriority.LOW
         }
-        if (HOSTS.contains(intent.data?.host)) {
+        if (SUPPORTED_HOSTS.contains(intent.data?.host)) {
             return DeepLinkParserDelegatePriority.HIGH
         }
         return DeepLinkParserDelegatePriority.LAST_CHANCE
@@ -38,6 +37,6 @@ class DeepLinkWebUpdater @Inject constructor() : DeepLinkParserDelegate, LogTagP
         val link = uri.getQueryParameter(QUERY_URL) ?: return null
         val version = uri.getQueryParameter(QUERY_VERSION) ?: ""
         val channel = uri.getQueryParameter(QUERY_CHANNEL) ?: ""
-        return Deeplink.WebUpdate(link, "$channel $version", intent)
+        return Deeplink.WebUpdate(link, "$channel $version")
     }
 }