+++ components/firstpair/impl/src/main/java/com/flipperdevices/firstpair/impl/composable/tos/ComposableTOS.kt
@@ -1,6 +1,7 @@
 package com.flipperdevices.firstpair.impl.composable.tos
 
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -16,6 +17,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.core.markdown.ClickableUrlText
 import com.flipperdevices.core.ui.ktx.ComposableFlipperButton
+import com.flipperdevices.core.ui.ktx.SetUpNavigationBarColor
+import com.flipperdevices.core.ui.ktx.SetUpStatusBarColor
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.firstpair.impl.R
@@ -23,10 +26,15 @@ import com.flipperdevices.core.ui.res.R as DesignSystem
 
 @Composable
 fun ComposableTOS(onApplyPress: () -> Unit, modifier: Modifier = Modifier) {
-    Column(modifier = modifier) {
+    SetUpStatusBarColor(LocalPallet.current.background)
+    Column(
+        modifier = modifier
+            .background(LocalPallet.current.background)
+    ) {
         ComposableTutorial(modifier = Modifier.weight(weight = 1f))
         ComposableFooter(onApplyPress)
     }
+    SetUpNavigationBarColor(LocalPallet.current.background)
 }
 
 @Composable