+++ components/info/impl/src/main/java/com/flipperdevices/info/impl/api/InfoFeatureEntryImpl.kt
@@ -1,15 +1,18 @@
 package com.flipperdevices.info.impl.api
 
-import android.content.Intent
+import android.net.Uri
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
+import androidx.navigation.navArgument
 import androidx.navigation.navDeepLink
 import androidx.navigation.navigation
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.ui.navigation.AggregateFeatureEntry
 import com.flipperdevices.core.ui.navigation.LocalGlobalNavigationNavStack
+import com.flipperdevices.deeplink.model.Deeplink
 import com.flipperdevices.deeplink.model.DeeplinkConstants
+import com.flipperdevices.deeplink.model.DeeplinkNavType
 import com.flipperdevices.info.api.screen.InfoFeatureEntry
 import com.flipperdevices.info.impl.compose.screens.ComposableDeviceInfoScreen
 import com.flipperdevices.info.impl.compose.screens.ComposableFullDeviceInfoScreen
@@ -18,7 +21,13 @@ import com.flipperdevices.updater.api.UpdaterCardApi
 import com.flipperdevices.updater.api.UpdaterFeatureEntry
 import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.anvil.annotations.ContributesMultibinding
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
 import javax.inject.Inject
+
+private const val DEEPLINK_KEY = DeeplinkConstants.KEY
+private const val DEEPLINK_SCHEME = DeeplinkConstants.SCHEMA
+private const val DEEPLINK_WEB_UPDATER_URL = "${DEEPLINK_SCHEME}web_updater={$DEEPLINK_KEY}"
 
 @ContributesBinding(AppGraph::class, InfoFeatureEntry::class)
 @ContributesMultibinding(AppGraph::class, AggregateFeatureEntry::class)
@@ -28,15 +37,32 @@ class InfoFeatureEntryImpl @Inject constructor(
     private val updaterFeatureEntry: UpdaterFeatureEntry
 ) : InfoFeatureEntry {
 
-    private val deeplinkKey = "{${DeeplinkConstants.KEY}}"
-
-    private fun start(): String = "@${ROUTE.name}"
+    private val infoRoute = "@${ROUTE.name}/$DEEPLINK_KEY={$DEEPLINK_KEY}"
 
     override fun fullInfo(): String = "@${ROUTE.name}full"
+    override fun getWebUpdateByDeeplink(deeplink: Deeplink): String {
+        val deeplinkStr = Uri.encode(Json.encodeToString(deeplink))
+        return "${DEEPLINK_SCHEME}web_updater=$deeplinkStr"
+    }
+
+    private val arguments = listOf(
+        navArgument(DeeplinkConstants.KEY) {
+            nullable = true
+            type = DeeplinkNavType()
+        }
+    )
+
+    private val deeplinkArguments = listOf(
+        navDeepLink { uriPattern = DEEPLINK_WEB_UPDATER_URL }
+    )
 
     override fun NavGraphBuilder.navigation(navController: NavHostController) {
-        navigation(startDestination = start(), route = ROUTE.name) {
-            composable("@${ROUTE.name}") {
+        navigation(startDestination = infoRoute, route = ROUTE.name) {
+            composable(
+                route = infoRoute,
+                arguments = arguments,
+                deepLinks = deeplinkArguments
+            ) {
                 val globalNavController = LocalGlobalNavigationNavStack.current
                 ComposableDeviceInfoScreen(
                     updaterCardApi,
@@ -49,29 +75,6 @@ class InfoFeatureEntryImpl @Inject constructor(
             }
             composable("@${ROUTE.name}full") {
                 ComposableFullDeviceInfoScreen(navController)
-            }
-            composable(
-                route = "@${ROUTE.name}/$deeplinkKey",
-                deepLinks = listOf(
-                    navDeepLink {
-                        uriPattern = "${DeeplinkConstants.LAB_FLIPPER_NET}/$deeplinkKey"
-                        action = Intent.ACTION_VIEW
-                    },
-                    navDeepLink {
-                        uriPattern = "${DeeplinkConstants.MY_FLIPPER_DEV}/$deeplinkKey"
-                        action = Intent.ACTION_VIEW
-                    }
-                )
-            ) {
-                val globalNavController = LocalGlobalNavigationNavStack.current
-                ComposableDeviceInfoScreen(
-                    updaterCardApi,
-                    onOpenFullDeviceInfo = { navController.navigate(fullInfo()) },
-                    onOpenOptions = { navController.navigate(settingFeatureEntry.ROUTE.name) },
-                    onStartUpdateRequest = {
-                        globalNavController.navigate(updaterFeatureEntry.getUpdaterScreen(it))
-                    }
-                )
             }
         }
     }