+++ components/singleactivity/impl/src/main/java/com/flipperdevices/singleactivity/impl/SingleActivity.kt
@@ -85,7 +85,7 @@ class SingleActivity :
                 ) {
                     ComposableSingleActivityNavHost(
                         navController = navControllerLocal,
-                        bottomNavigationFeatureEntry = bottomNavigationFeatureEntry,
+                        starDestination = deepLinkHelper.getStartDestination(),
                         featureEntries = featureEntriesMutable.toPersistentSet(),
                         composableEntries = composableEntriesMutable.toPersistentSet(),
                         modifier = Modifier