+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - [FIX] Firstpair flow open by navigation graph route
 - [FIX] Fix crush with inapp storage on startup
 - [FIX] Fix change color scheme bug
+- [FIX] Fix status bar color
 - [REFACTOR] Migrate bottom bar to compose navigation
 - [REFACTOR] Bump Android Gradle Plugin
 - [REFACTOR] Fix detekt compose issues
@@ -33,7 +34,7 @@
 - [REFACTOR] Enable detekt formatting with rules
 - [REFACTOR] Fix gradle warnings (Deprecated API, PredefinedEnumMap)
 - [REFACTOR] Replace some clickable to card function + reuse orange app bar
-- [REFACTOR] Google Compose Navigation (archive, firstpair, widgets(composable entry), updater, import)
+- [REFACTOR] Google Compose Navigation (archive, firstpair, widgets, updater, import, deeplinks)
 - [REFACTOR] Migrate to new device info api
 - [REFACTOR] Migrate to isSupported in version api